+++ plugin/McpLink/Source/McpLinkEditor/Private/McpLinkEditorRoutes.h
@@ -13,6 +13,7 @@ namespace McpLink
 	void RegisterEditorOpsRoutes(FMcpLinkCoreModule& Core);
 	void RegisterPerfRoutes(FMcpLinkCoreModule& Core);
 	void RegisterVisualLogRoutes(FMcpLinkCoreModule& Core);
+	void RegisterUiOpsRoutes(FMcpLinkCoreModule& Core);
 	void RegisterPieRoutes(FMcpLinkCoreModule& Core);
 	void RegisterCaptureRoutes(FMcpLinkCoreModule& Core);
 	void RegisterIntrospectRoutes(FMcpLinkCoreModule& Core);